requirements.txt
# requirements.txt

# --- Core de Scraping e Análise ---
requests
tldextract
beautifulsoup4
lxml
faker

# --- Módulos de Comunicação C2 ---
paho-mqtt
websocket-client
PySocks
